favicon fails saved to file, refactor for discrete function
